chore(typescript): migrate input.ts to typescript
